+++ src/FreeElephants/Validation/Rule/Boolean.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace FreeElephants\Validation\Rule;
+
+use Rakit\Validation\Rule;
+
+class Boolean extends Rule
+{
+    public function check($value): bool
+    {
+        return is_bool($value);
+    }
+}