+++ src/FreeElephants/Validation/Rules.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace FreeElephants\Validation;
+
+interface Rules
+{
+    public function getRules(): array;
+}